+++ index.js
@@ -1,4 +1,5 @@
 // Third-party dependencies
+import "react-native-gesture-handler" // This must be at the top of the file. See https://reactnavigation.org/docs/en/next/getting-started.html
 import {
     AppRegistry
 } from 'react-native'